==> notebooks/01_data_audit.ipynb <==
# Cafe Sales Analytics Projects

## Objectives
Audit the raw datasets of Cafe Sales to understand the structure, quality issues and potential cleaning steps to take.

import pandas as pd
import numpy as np

import os
os.getcwd()

#import CSV
data_path = "../data/raw/dirty_cafe_sales.csv"
try:
    df = pd.read_csv(data_path)
except FileNotFoundError:
    print("ERROR : File Not Found")
pd.reset_option("display.max_row")

df

## Check Structure

df.shape

df.columns

df.info()

## Sample Data

df.head(10)

df.tail(10)

## Check Data Quality

df.isnull().sum()

(df.isnull().sum()/ len(df))*100

df.duplicated().sum()

## Distribution Analysis

df.describe(include="all")

## Initial Observations
- This Dataset have 10 000 rows with 8 columns 
- no duplicate data
- unique transaction id with no missing or invalid data
- all column have the type object (needs to be fixed)

## Potential Issues
- there is quite a few null values per column as a percentage 
    Item                 3.33%
    Quantity             1.38%
    Price Per Unit       1.79%
    Total Spent          1.73%
    Payment Method      25.79%
    Location            32.65%
    Transaction Date     1.59%
- There is also a bunch of unknown and error values that need to be fixed for each colmn.
- missing values in the price per unit column needs to be verifyed and correct to the propre amount
- verify total spent calculations using Quantity * Price Per Unit.
- determining missing quantity by reverse calculating using the total and price per unit
- set transaction dates to datetime and order them
- fix data types for each columns

## Questions to Resolves
-TODO
