==> notebooks/02_data_cleaning.ipynb <==
# Cafe Sales Analytics Projects

## Objectives
Clean the raw datasets  Cafe Sales.

### SETUP for Cleaning

import pandas as pd
import numpy as np

#import CSV
data_path = "../data/raw/dirty_cafe_sales.csv"
try:
    df = pd.read_csv(data_path)
except FileNotFoundError:
    print("ERROR : File Not Found")
    

# see columns
df.columns

#make a copy to clean
clean_df = df.copy()
pd.set_option("display.max_row", None)
# check all unique values in the item colunm
clean_df["Item"].unique() 

### Clean 'Item' column

# change the type to str and strip leading an trailing spaces (not needed)
clean_df["Item"] = clean_df["Item"].astype('string').str.strip()
# set title format for item (not needed)
clean_df["Item"] = clean_df["Item"].str.title()
#check changes
clean_df["Item"].unique()

clean_df["Item"].dtype

# set erroneous values in the item column to NULL
prob = ['UNKNOWN','ERROR','nan','Unknown','Nan','Error','<Na>']
clean_df["Item"] = clean_df["Item"].replace(prob,pd.NA)
#check changes
clean_df["Item"].unique()

### Clean 'Quantity' column

# check unique values
clean_df['Quantity'].unique()

# remove all teh problem values from Quantity column 
clean_df['Quantity'] = clean_df['Quantity'].replace(prob,pd.NA)
# set column type to int32
clean_df['Quantity'] = clean_df['Quantity'].astype('Int32')
# verify changes
clean_df['Quantity'].unique()

### Clean 'Price Per Unit' Column

# replace erroneous values in the Price Per Unit column to NULL
clean_df['Price Per Unit'] = clean_df['Price Per Unit'].replace(prob,np.nan)
# set the type to float
clean_df['Price Per Unit'] = pd.to_numeric(clean_df['Price Per Unit'], errors = "coerce")
# verify type
clean_df['Price Per Unit'].dtype

# verify correct values
clean_df['Price Per Unit'].unique()

### Clean 'Total Spent' Column

# check out the values in the column
clean_df['Total Spent'].unique()

# remove erroneous values in the column
clean_df['Total Spent'] = clean_df['Total Spent'].replace(prob,np.nan)
# update the type to a float while maintening NULL values
clean_df['Total Spent'] = pd.to_numeric(clean_df['Total Spent'],errors="coerce")
#check changed type
clean_df['Total Spent'].dtype

# recheck column values
clean_df['Total Spent'].unique()

#check columns (to view them easier)
df.columns

### Clean 'Payment Method' Column

# view unique values to clean
clean_df['Payment Method'].unique()

#Set type and strip trailing and leading (not needed)
clean_df['Payment Method'] = clean_df['Payment Method'].astype("string").str.strip()
# set title format(not needed)
clean_df['Payment Method'] = clean_df['Payment Method'].str.title()

# remove invalid values and replace with NULL
clean_df['Payment Method'] = clean_df['Payment Method'].replace(prob,pd.NA)
# check changes and type
clean_df['Payment Method'].unique()

### Clean 'Location' Column

# check what needs to be corrected
clean_df['Location'].unique()

# set type to string
clean_df['Location'] = clean_df['Location'].astype("string")
# remove invalid values and replace with NULL
clean_df['Location'] = clean_df['Location'].replace(prob,pd.NA)

# check if corrected
clean_df['Location'].unique()

### Clean 'Transaction Date' Column

# remove invalid values and replace with NULL
clean_df['Transaction Date'] = clean_df['Transaction Date'].replace(prob,pd.NA)

# convert to datetime type
clean_df['Transaction Date'] = pd.to_datetime(clean_df['Transaction Date'])
# check type
print(clean_df['Transaction Date'].dtype)

### SAVE the Cleaned dataset to a new CSV

clean_df.to_csv("../data/processed/cleaned_cafe_sales_without_filling.csv", index = False)
