==> notebooks/03_data_filling.ipynb <==
# Cafe Sales Analytics Projects

## Objectives
fill the missing values in the cleaned datasets Cafe Sales.

### SETUP for Filling values

import pandas as pd
import numpy as np

#import CSV
data_path = "../data/processed/cleaned_cafe_sales_without_filling.csv"
try:
    df = pd.read_csv(data_path,
                     dtype= {
                        "Transaction ID" : "object",
                        "Item" : "string",
                        "Quantity": "Int32",
                        "Price Per Unit": "float64",
                        "Total Spent": "float64",
                        "Payment Method": "string", 
                        "Location": "string", 
                     } 
                     ,parse_dates=["Transaction Date"]
                     )
except FileNotFoundError:
    print("ERROR : File Not Found")

#make a copy to fill
fill_df = df.copy()
pd.set_option("display.max_row", None)

fill_df.columns

## Fill missing values in 'Item' 
Missing Item names were inferred only using the price per unit column and only when an item's price per unit is unique.
items that have the same prices where left out to avoid assumptions.

From 969 missing values in the item columns to 480 missing values. I was able to infer a total of 489 item names first based on the Price per unit column followed by dividing the total spent column with the quantity column to figure out the price per unit and therefore the missing item values. 

print(df["Item"].isnull().sum())

#find known items
item_known = fill_df[fill_df["Item"].notna()]

# find items with unique price per unit
price_count = item_known.groupby("Price Per Unit")["Item"].nunique()
print(price_count)

# get safe items to infer name from the price per unit == 1
safe_price = price_count[price_count == 1].index
safe_price

# create a lookup table
price_to_item = (item_known[item_known["Price Per Unit"].isin(safe_price)]
                 .drop_duplicates("Price Per Unit")
                 .set_index("Price Per Unit")["Item"])
print(price_to_item)

# find missing values in the item column
mask = fill_df["Item"].isna()
# fill missing values using the mask
fill_df.loc[mask,"Item"] = ( fill_df.loc[mask,"Price Per Unit"].map(price_to_item) )

print(fill_df["Item"].isnull().sum())

### Fill missing values in 'Item' Using Total Spent and Quantity

Missing Item names were inferred using the Total Spent and Quantity column to get the price per unit.

# create a temp df for calculated price
fill_df["Computed_price"] = (fill_df["Total Spent"]/ fill_df["Quantity"]).round(2)
fill_df.columns

# find missing values in the item column
mask = (fill_df["Item"].isna() 
        &  
        fill_df["Computed_price"].isin(safe_price)
        )
# fill missing values using the mask
fill_df.loc[mask,"Item"] = ( fill_df.loc[mask,"Computed_price"].map(price_to_item) )

print(fill_df["Item"].isnull().sum())

## Fill missing values in 'Price per unit' 

Missing values in price per unit were infered using the item column and the computation of total spent/quantity columns

From 533 missing values in the price per unit column to 6 missing values. These values were infered by first using the item column and the using the computation total spent divided by quantity to fill the remaining values if both values were not null and valid

print(fill_df["Price Per Unit"].isna().sum())

# find all not Null values in Item and Price per unit
valid = fill_df[fill_df["Item"].notna() & fill_df["Price Per Unit"].notna()]

# find out if each item has a unique price 
price_count = valid.groupby("Item")["Price Per Unit"].nunique()
price_count

# get items that sell at one price (safe)
safe_item = price_count[price_count == 1].index
safe_item

#create a lookup table item -> prices
item_to_price = (
    valid[valid["Item"].isin(safe_item)]
    .drop_duplicates("Item")
    .set_index("Item")["Price Per Unit"]
)
print(item_to_price)

# fill missing values in price per unit
mask = (fill_df["Price Per Unit"].isna() 
        &  
        fill_df["Item"].isin(item_to_price.index)
        )
# fill missing values using the mask
fill_df.loc[mask,"Price Per Unit"] = ( fill_df.loc[mask,"Item"].map(item_to_price) )

print(f"Null values in Price Per Unit: {fill_df['Price Per Unit'].isna().sum()},\nNull values in Item : {fill_df['Item'].isna().sum()}")

### Fill in missing values in price per unit using total spent/quantity

Remaining missing values were infered using total spent divided by quantity if both values were not null and valid



#create a mask for safe rows
mask = (
    fill_df["Price Per Unit"].isna() &
    fill_df["Total Spent"].notna() &
    fill_df["Quantity"].notna() &
    (fill_df["Quantity"] > 0)
    )

# fill missing values in price per unit
fill_df.loc[mask,"Price Per Unit"] = ( fill_df.loc[mask,"Computed_price"].round(2) )

fill_df.drop(columns="Computed_price", inplace= True)
print(fill_df['Price Per Unit'].isna().sum())

## Fill missing values in 'Quantity' 
Missing values in Quantity were infered using the computation of total spent/price per unit columns

From 479 missing values in the quantity column to 23 NaN. This was achieved by infering the quantity using the total spent divided by the price per unit when both values were not null and price per unit is greater than zero. Also making sure the computation was rounded to be an integer. Only affected the quantity values that were null and fufilled the aformention conditions.

print(fill_df["Quantity"].isna().sum())

# create a temp df for calculated quantity
fill_df["Computed_quantity"] = (fill_df["Total Spent"]/ fill_df["Price Per Unit"])

fill_df.columns

# create a mask for the proper conditions
mask = (
    fill_df["Quantity"].isna() &
    fill_df["Total Spent"].notna() &
    fill_df["Price Per Unit"].notna() &
    (fill_df["Price Per Unit"] > 0) &
    (fill_df["Computed_quantity"].round() == fill_df["Computed_quantity"])
)

#fill in missing Quantities
fill_df.loc[mask,"Quantity"] = ( fill_df.loc[mask,"Computed_quantity"]).astype(int)

#check changes
print(fill_df["Quantity"].isna().sum())

fill_df.drop(columns= "Computed_quantity",inplace= True)
fill_df.columns

fill_df.loc[fill_df["Quantity"].isna() ,["Item", "Quantity", "Price Per Unit", "Total Spent"]].head(23)

## Fill missing values in 'Total Spent' 
Missing values in Total Spent were infered using the computation of Quantity * price per unit columns

From 503 missing values in the total spent column to 23 NaN. This was achieved by infering the total spent using the quantity  multiplied by the price per unit when both values were not null and greater than zero. Also making sure the computation was rounded to 2 decimal. Only affected the total spent values that were null and fufilled the aformention conditions.

print(fill_df["Total Spent"].isna().sum())

# create a temp df for calculated total spent
fill_df["Computed_total"] = (fill_df["Quantity"] * fill_df["Price Per Unit"])

# create a mask to check all conditions
mask = (
    fill_df["Total Spent"].isna() &
    fill_df["Quantity"].notna() &
    fill_df["Price Per Unit"].notna() &
    (fill_df["Quantity"] > 0) &
    (fill_df["Price Per Unit"] > 0)
)

# fill in the null values in total spent 
fill_df.loc[mask, "Total Spent"] = (fill_df.loc[mask, "Computed_total"].round(2))

# check changes
print(fill_df["Total Spent"].isna().sum())

fill_df.drop(columns= "Computed_total", inplace = True)

fill_df.columns

## Save the infered dataset to a new CSV

fill_df.to_csv("../data/processed/infered_cafe_sales_completed.csv",index = False)
