==> notebooks/04_data_feature.ipynb <==
# Cafe Sales Analytics Projects

## Objectives
Create useful columns to make analyse of the cafe sales data easier to answer business related question easier.

### SETUP for Feature Engineering

import pandas as pd
import numpy as np

# import CSV
data_path = "../data/processed/infered_cafe_sales_completed.csv"
try:
    df = pd.read_csv(data_path,
                     dtype= {
                        "Transaction ID" : "object",
                        "Item" : "string",
                        "Quantity": "Int32",
                        "Price Per Unit": "float64",
                        "Total Spent": "float64",
                        "Payment Method": "string", 
                        "Location": "string", 
                     } 
                     ,parse_dates=["Transaction Date"]
                     )
except FileNotFoundError:
    print("ERROR : File Not Found")

#make a copy just in case
feature_df = df.copy()
pd.set_option("display.max_row", None)

feature_df.columns

feature_df.info()

## Time based features

#month features
feature_df["Month"] = feature_df["Transaction Date"].dt.month.astype("Int64")

# year freatures
feature_df["Week"] = feature_df["Transaction Date"].dt.isocalendar().week.astype("Int64")

#day features
feature_df["Day Name"] = feature_df["Transaction Date"].dt.day_name()
feature_df["Day Of Week"] =feature_df["Transaction Date"].dt.dayofweek

# check weekend
feature_df["Is Weekend"] = feature_df["Day Of Week"].isin([5,6]).astype(bool)

# check changes
feature_df.drop(columns="Day Of Week",inplace = True)


feature_df.sample()

## Transaction features

# check if math is correct
check = (feature_df["Quantity"]*feature_df["Price Per Unit"]).round(2)
feature_df["price_check"] = (check != feature_df["Total Spent"].round(2)).astype("Int64")

feature_df["price_check"].value_counts()

feature_df.drop(columns= "price_check",inplace = True)

feature_df.columns

### Transaction Size

feature_df["Transaction Size"] = pd.cut(
    feature_df["Total Spent"],
    bins = [0,10,20,float("inf")],
    labels=["Small","Medium","Large"],
    include_lowest=True
)

feature_df[["Total Spent","Transaction Size"]].sample(10)

feature_df["Transaction Size"].value_counts(dropna=False)

### Bulk Orders

feature_df["Bulk Order"] = pd.NA
feature_df.loc[feature_df["Quantity"] > 1, "Bulk Order"] = 1 
feature_df.loc[feature_df["Quantity"] == 1, "Bulk Order"] = 0 
feature_df["Bulk Order"] = feature_df["Bulk Order"].astype("Int64")

feature_df[["Quantity","Bulk Order"]].sample(10)

feature_df["Bulk Order"].value_counts(dropna=False)

feature_df["Bulk Order"].dtype

feature_df.sample(5)

## Save Final Dataset to CSV

feature_df.to_csv("../data/processed/final_cafe_sales.csv",na_rep="\\N",index = False)

feature_df.info()
